enhance logging for HTTP requests/responses
