+++ RechargeSharp.v2021-11/Configuration/DependencyInjection/RechargeSharpDependencyInjection.cs
@@ -1,10 +1,13 @@
 using System.Net.Http.Headers;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+using Microsoft.Extensions.Http;
 using RechargeSharp.v2021_11.Endpoints.Addresses;
 using RechargeSharp.v2021_11.Endpoints.Customers;
 using RechargeSharp.v2021_11.Endpoints.PaymentMethods;
 using RechargeSharp.v2021_11.Endpoints.Subscriptions;
 using RechargeSharp.v2021_11.Utilities;
+using RechargeSharp.v2021_11.Utilities.Logging;
 
 namespace RechargeSharp.v2021_11.Configuration.DependencyInjection;
 
@@ -25,8 +28,9 @@ public static class RechargeSharpDependencyInjection
             opts.BaseAddress = new Uri("https://api.rechargeapps.com/");
         }).AddHttpMessageHandler<LoggingHttpHandler>();
 
-        services.AddTransient(x => options);
         services.AddLogging();
+        // This removes the default logging behaviour in the HttpClient, that logs all requests and responses in the Information log level
+        services.RemoveAll<IHttpMessageHandlerBuilderFilter>();
 
         services
             .AddTransient<LoggingHttpHandler>()
