+++ RechargeSharp.v2021-11/RechargeSharp.v2021-11.csproj
@@ -16,8 +16,4 @@
       <PackageReference Include="Polly" Version="7.2.3" />
     </ItemGroup>
 
-    <ItemGroup>
-      <Folder Include="Utilities\Logging" />
-    </ItemGroup>
-
 </Project>
