+++ RechargeSharp.v2021-11/Utilities/Logging/LoggingHttpHandlerOptions.cs
@@ -1,0 +1,8 @@
+using Microsoft.Extensions.Logging;
+
+namespace RechargeSharp.v2021_11.Utilities.Logging;
+
+public class LoggingHttpHandlerOptions
+{
+    public LogLevel LogLevel { get; set; } = LogLevel.Debug;
+}
